Move transaction signer module out of transaction.rb
